исправлены все замечания

+++ src/test/java/tests/AuthorizedBaseTest.java
@@ -3,8 +3,12 @@ package tests;
 import api.UserApiClient;
 import data.UserData;
 import data.UserDataFactory;
+import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
+
+import static org.apache.http.HttpStatus.SC_OK;
+import static org.junit.Assert.assertEquals;
 
 public class AuthorizedBaseTest extends BaseTest {
     protected UserData testUser;
@@ -13,8 +17,13 @@ public class AuthorizedBaseTest extends BaseTest {
     @Before
     public void initAuthorizedUser() {
         testUser = UserDataFactory.createValidUser();
-        UserApiClient.createUser(testUser);
-        authToken = UserApiClient.getAuthToken(testUser);
+
+        Response createResponse = UserApiClient.createUser(testUser);
+        assertEquals(SC_OK, createResponse.statusCode());
+
+        Response loginResponse = UserApiClient.loginUser(testUser);
+        assertEquals(SC_OK, loginResponse.statusCode());
+        authToken = loginResponse.path("accessToken");
 
         mainPage.clickSignInBtn();
         authPage.typeEmail(testUser.getEmail());
@@ -25,6 +34,8 @@ public class AuthorizedBaseTest extends BaseTest {
 
     @After
     public void cleanupUser() {
-        UserApiClient.deleteUser(authToken);
+        if (authToken != null && !authToken.isEmpty()) {
+            UserApiClient.deleteUser(authToken);
+        }
     }
 }
